+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.7.0]
+### Added
+- `yarn build` results in multiple outputs for each component instead of one single huge file
+
 ### Refactored
 - Import fa_icons.css in all components using `fa-icon`